# docs/swagger.yaml
basePath: /api
definitions:
  dto.ErrorResponse:
    properties:
      code:
        description: Code is the error code that identifies the specific error type
        type: integer
      details:
        description: Details is an optional field for providing additional context
          information
      message:
        description: Message is a user-friendly error description
        type: string
    type: object
  dto.UserLoginReq:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.UserLoginResp:
    properties:
      accessToken:
        type: string
    type: object
  dto.UserResp:
    properties:
      create_time:
        type: string
      email:
        type: string
      id:
        type: integer
      update_time:
        type: string
      username:
        type: string
    type: object
host: 127.0.0.1:8000
info:
  contact: {}
  description: This is an API for Cinemago
  title: Cinemago
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: User login with username and password returns an access token
      parameters:
      - description: LoginInfo
        in: body
        name: loginReq
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserLoginResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: user login
      tags:
      - Auth
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Get user details by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get user by ID
      tags:
      - User
swagger: "2.0"